requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.2
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
